+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All notable changes to this project will be documented in this file.
 
 - `allow_latest_inconsistent_data` option in paginated `blockchain` queries:
   > By default there a delay of realtime data to ensure impossibility of data inserts before latest accessible cursor. It is possible to disable this guarantee and to reduce delay of realtime data by setting this flag to true.
+- master config `p30`, `p40`, `p42` fields types
+- `prev_code_hash` account field
 
 ### Fixed